requirements.txt
aiofiles==0.7.0
appdirs==1.4.3
APScheduler==3.7.0
arrow==1.1.1
attrs==21.2.0
bcrypt==3.2.0
bitarray==2.1.3
boto3==1.17.102
botocore==1.20.102
CacheControl==0.12.6
cached-property==1.5.2
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.4.7
decorator==5.0.9
distlib==0.3.0
distro==1.4.0
ecdsa==0.14.1
et-xmlfile==1.1.0
fastapi==0.65.1
frictionless==4.11.0
frictionless-geojson==0.0.2
greenlet==1.1.0
h11==0.12.0
html5lib==1.0.1
httptools==0.1.2
idna==2.8
ijson==3.1.4
iniconfig==1.1.1
ipaddr==2.2.0
isodate==0.6.0
jmespath==0.10.0
json-table-schema==0.2.1
jsonlines==2.0.0
jsonpointer==2.1
jsonschema==3.2.0
linear-tsv==1.1.0
lockfile==0.12.2
lxml==4.6.3
messytables==0.15.2
msgpack==0.6.2
mysql==0.0.2
mysqlclient==2.0.3
numpy==1.21.0
openpyxl==3.0.7
packaging==20.3
pandas==1.3.0
passlib==1.7.4
pep517==0.8.2
petl==1.7.4
pluggy==0.13.1
progress==1.5
py==1.10.0
pyasn1==0.4.8
pybloom-live==3.1.0
pycparser==2.20
pydantic==1.8.2
pymongo==3.11.4
pyparsing==2.4.6
pyrsistent==0.17.3
pytest==6.2.4
python-dateutil==2.8.1
python-dotenv==0.17.1
python-jose==3.2.0
python-magic==0.4.24
python-multipart==0.0.5
python-slugify==5.0.2
python3-gearman==0.1.0
pytoml==0.1.21
pytz==2021.1
PyYAML==5.4.1
requests==2.22.0
retrying==1.3.3
rfc3986==1.5.0
rsa==4.7.2
s3transfer==0.4.2
Shapely==1.7.1
shellingham==1.4.0
simpleeval==0.9.10
six==1.14.0
SQLAlchemy==1.4.15
starlette==0.14.2
starlette-context==0.3.2
stringcase==1.2.0
tabulator==1.53.5
text-unidecode==1.3
toml==0.10.2
typer==0.3.2
typing-extensions==3.10.0.0
tzlocal==2.1
unicodecsv==0.14.1
urllib3==1.25.8
uvicorn==0.13.4
uvloop==0.15.2
validators==0.18.2
watchgod==0.7
webencodings==0.5.1
websockets==8.1
xlrd==2.0.1
